refactor: switch from email as primary key to id because of postman tests. In process of fixing all feature: add UserDbStorage
