+++ src/main/java/ru/yandex/practicum/filmorate/services/Likeable.java
@@ -5,8 +5,8 @@ import ru.yandex.practicum.filmorate.model.Film;
 import java.util.List;
 
 public interface Likeable {
-    void addLike(int filmID, int userID);
-    void removeLike(int filmID, int userID);
+    void addLike(int filmID, String userEmail);
+    void removeLike(int filmID, String userEmail);
     List<Film> getMostPopularFilms(int topFilmsAmount);
-    void checkFilmAndUserPresence(int filmID, int userID);
+    void checkFilmAndUserPresence(int filmID, String userEmail);
 }
