+++ src/main/resources/application.properties
@@ -1,7 +1,6 @@
 logging.level.ru.yandex.practicum.filmorate=trace;
 
 spring.sql.init.mode=always
-# ? jdbc-url ???????, ??? ?????? ????? ????????? ? ????
 spring.datasource.url=jdbc:h2:file:./db/filmorate
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
